09_DJANGO_LMS/requirements.txt
crispy-bootstrap5==0.6
Django==4.0.5
django-debug-toolbar==3.4.0
django-extensions==3.1.5
django-filter==21.1
Faker==13.13.0
flake8==4.0.1
flake8-builtins==1.5.3
flake8-import-order==0.18.1
flake8-polyfill==1.0.2
ipython==8.4.0
pep8-naming==0.13.0
Pillow==9.1.1
python-dotenv==0.20.0
webargs==8.1.0
